Storing info in room database (not happening automatically yet)

+++ app/src/main/java/com/example/fit5046_lab4_group3_ass2/room/DayUse.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class DayUse(
     @PrimaryKey(autoGenerate = true)
     val uid: Int = 0,
-    val date: Int,
-    val use: Int,
+    val date: Long,
+    val use: Float,
     val price: Int
 )
